Move predicates into PbftLog
